+++ src/flow_control_test/test4.cpp
@@ -9,9 +9,9 @@ class Test4_ActionParameterGetter : public FlowControl::IActionParameterGetter
 public:
 	state::CardRef GetDefender(std::vector<state::CardRef> const& targets)
 	{
-		if (next_defender_count >= 0) assert(next_defender_count == targets.size());
+		if (next_defender_count >= 0) assert(next_defender_count == (int)targets.size());
 		assert(next_defender_idx >= 0);
-		assert(next_defender_idx < targets.size());
+		assert(next_defender_idx < (int)targets.size());
 		return targets[next_defender_idx];
 	}
 
@@ -22,7 +22,7 @@ public:
 
 	state::CardRef GetSpecifiedTarget(state::State & state, state::CardRef card_ref, std::vector<state::CardRef> const& targets)
 	{
-		assert(targets.size() == next_specified_target_count);
+		assert((int)targets.size() == next_specified_target_count);
 
 		if (next_specified_target_idx < 0) return state::CardRef();
 		else return targets[next_specified_target_idx];
@@ -30,7 +30,7 @@ public:
 
 	Cards::CardId ChooseOne(std::vector<Cards::CardId> const& cards) {
 		choose_one_called = true;
-		assert(next_choose_one_count == cards.size());
+		assert(next_choose_one_count == (int)cards.size());
 		return cards[next_choose_one_idx];
 	}
 
@@ -116,7 +116,7 @@ static state::CardRef AddHandCard(Cards::CardId id, FlowControl::FlowContext & f
 	}
 	else {
 		++hand_count;
-		assert(state.GetBoard().Get(player).hand_.Size() == hand_count);
+		assert((int)state.GetBoard().Get(player).hand_.Size() == hand_count);
 		assert(state.GetBoard().Get(player).hand_.Get(hand_count - 1) == ref);
 		assert(state.GetCardsManager().Get(ref).GetZone() == state::kCardZoneHand);
 		assert(state.GetCardsManager().Get(ref).GetZonePosition() == (hand_count - 1));
